+++ CHANGELOG.md
@@ -13,12 +13,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Hardware
 
+## [25.8.26]
+
+### Added
+
+### Changed
+- Unique (static) name generation for Android devices to prevent re-pairing issues.
+
+### Hardware
+
+
+## [25.8.18]
+
+### Added
+
+### Changed
+- Added better BLE device logging, even when a connection isn't made.
+- Fixed Echelon connections. 
+
+### Hardware
+
+## [25.8.16]
+
+### Added
+
+### Changed
+
+### Hardware
 
 ## [25.8.3]
 
 ### Added
 
 ### Changed
+- Fixed rare crash due to calling new scan before the previous onScanEnd callback was complete. 
+- Removed serial printf's during updates as it was causing occasional update issues.
 
 ### Hardware
 
@@ -35,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [25.7.29]
 
 ### Added
+- Added state machine for shifters allowing faster and more reliable shifting. 
 
 ### Changed
 - Removed unused file. 
@@ -45,6 +75,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed edge cases of pTab4Pwr causing runaways.
 - Reduced the power output when no power table.
 - Fixed Dircon with Mywoosh.
+- Fixed Bug with Peloton data being requested too often. 
+- Fixed Bug with Dircon data being sent too often. 
+- Fixed Bug with BLE TX/RX being overwhelmed by the above two. 
+- Homing now takes multiple samples at the start of the run. 
 
 ### Hardware